+++ Sources/BuilderIO/Helpers/CSSAlignments.swift
@@ -1,0 +1,54 @@
+import SwiftUICore
+
+class CSSAlignments {
+
+  static func textAlignment(responsiveStyles: [String: String]) -> HorizontalAlignment {
+    if let textAlign = responsiveStyles["textAlign"] {
+      switch textAlign {
+      case "center":
+        return .center
+      case "left", "start":  // "start" is also a common value in some contexts
+        return .leading
+      case "right", "end":  // "end" is also a common value
+        return .trailing
+      case "justify":
+        break  // Fall through to next checks
+      default:
+        break  // Unknown textAlign value, fall through
+      }
+    }
+
+    return .leading
+  }
+
+  static func horizontalAlignment(
+    marginsLeft: String?, marginsRight: String?, justify: String?, alignItems: String?,
+    responsiveStyles: [String: String]
+  ) -> HorizontalAlignment {
+
+    if (marginsLeft == "auto" && marginsRight == "auto") || justify == "center"
+      || alignItems == "center"
+    {
+      return .center
+    } else if marginsRight == "auto" || justify == "flex-start" || alignItems == "flex-start" {
+      return .leading
+    } else if marginsLeft == "auto" || justify == "flex-end" || alignItems == "flex-end" {
+      return .trailing
+    }
+
+    return textAlignment(responsiveStyles: responsiveStyles)
+  }
+
+  static func verticalAlignment(justify: String?, alignItems: String?) -> VerticalAlignment {
+
+    if justify == "center" || alignItems == "center" {
+      return .center
+    } else if justify == "flex-start" || alignItems == "flex-start" {
+      return .top
+    } else if justify == "flex-end" || alignItems == "flex-end" {
+      return .bottom
+    }
+    return .center
+  }
+
+}
